move D2Reader to reader.ts
